requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
faiss-cpu==1.10.0
fastapi==0.115.12
feedparser==6.0.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.48
langchain-text-splitters==0.3.7
langsmith==0.3.18
numpy==2.2.4
openai==1.68.2
orjson==3.10.16
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
